Revert "Merge branch 'master' into AvniConnector"
This reverts commit 12f8faaa1e247514ac5ca46b1b5209aa704cba4b, reversing
changes made to fbe32de1faa49b7a2716ee9dcf7b4df3a6bf92a3.

+++ airbyte-integrations/connectors/source-amazon-ads/source_amazon_ads/schemas/sponsored_brands.py
@@ -19,7 +19,7 @@ class BrandsCampaign(CatalogModel):
     state: str
     servingStatus: str
     brandEntityId: str
-    portfolioId: int
+    portfolioId: Decimal
     bidOptimization: bool = None
     bidMultiplier: Decimal = None
     adFormat: str
